+++ orcid-web/src/main/java/org/orcid/frontend/spring/ShibbolethAjaxAuthenticationSuccessHandler.java
@@ -25,51 +25,20 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Random;
 
-import javax.annotation.Resource;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.lang3.StringUtils;
 import org.orcid.core.exception.OrcidBadRequestException;
-import org.orcid.core.manager.InternalSSOManager;
-import org.orcid.core.manager.OrcidProfileManager;
-import org.orcid.core.manager.SourceManager;
-import org.orcid.core.manager.impl.OrcidUrlManager;
 import org.orcid.frontend.web.exception.FeatureDisabledException;
-import org.orcid.jaxb.model.message.Locale;
 import org.orcid.jaxb.model.message.OrcidProfile;
-import org.orcid.persistence.dao.ProfileDao;
-import org.orcid.persistence.dao.UserConnectionDao;
 import org.orcid.persistence.jpa.entities.UserconnectionEntity;
 import org.orcid.persistence.jpa.entities.UserconnectionPK;
-import org.orcid.utils.OrcidRequestUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import org.springframework.security.web.savedrequest.SavedRequest;
-import org.springframework.web.servlet.i18n.CookieLocaleResolver;
 
-public class ShibbolethAjaxAuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-
-    @Resource
-    private OrcidUrlManager orcidUrlManager;
-    
-    @Resource
-    private UserConnectionDao userConnectionDao;
-    
-    @Resource
-    private ProfileDao profileDao;
-    
-    @Resource
-    private InternalSSOManager internalSSOManager;
-    
-    @Resource
-    protected SourceManager sourceManager;
-    
-    @Resource
-    protected OrcidProfileManager orcidProfileManager;
+public class ShibbolethAjaxAuthenticationSuccessHandler extends AjaxAuthenticationSuccessHandlerBase {
     
     private static final String SHIB_IDENTITY_PROVIDER_HEADER = "shib-identity-provider";
     
@@ -106,18 +75,7 @@ public class ShibbolethAjaxAuthenticationSuccessHandler extends SimpleUrlAuthent
 		    userConnectionDao.persist(userConnectionEntity);
         }
     	
-    	String targetUrl = determineFullTargetUrlFromSavedRequest(request, response);
-        if (authentication != null) {
-            String orcidId = authentication.getName();
-            checkLocale(request, response, orcidId);
-            if(internalSSOManager.enableCookie()) {
-                internalSSOManager.writeCookie(orcidId, request, response);
-            }            
-            profileDao.updateIpAddress(orcidId, OrcidRequestUtil.getIpAddress(request));
-        }
-        if (targetUrl == null) {
-            targetUrl = determineFullTargetUrl(request, response);
-        }
+        String targetUrl = getTargetUrl(request, response, authentication);
         response.setContentType("application/json");
         response.getWriter().println("{\"success\": true, \"url\": \"" + targetUrl.replaceAll("^/", "") + "\"}");        
     }
@@ -128,72 +86,6 @@ public class ShibbolethAjaxAuthenticationSuccessHandler extends SimpleUrlAuthent
         }
     }
 
-    private String retrieveRemoteUser(Map<String, String> headers) {
-        for (String possibleHeader : POSSIBLE_REMOTE_USER_HEADERS) {
-            String userId = headers.get(possibleHeader);
-            if (userId != null) {
-                return userId;
-            }
-        }
-        throw new OrcidBadRequestException("Couldn't find remote user header");
-    }
-    
-    private OrcidProfile getRealProfile() {
-        String realOrcid = getRealUserOrcid();
-        return realOrcid == null ? null : orcidProfileManager.retrieveOrcidProfile(realOrcid);
-    }
-    
-    private String getRealUserOrcid() {
-        return sourceManager.retrieveRealUserOrcid();
-    }
-    
-    // new method - persist which ever local they logged in with
-    private void checkLocale(HttpServletRequest request, HttpServletResponse response, String orcidId) {
-        Locale lastKnownLocale = profileDao.retrieveLocale(orcidId);
-        if (lastKnownLocale != null) {
-
-            // have to read the cookie directly since spring has
-            // populated the request locale yet
-            CookieLocaleResolver clr = new CookieLocaleResolver();
-            // must match <property name="cookieName" value="locale_v3"
-            // />
-            clr.setCookieName("locale_v3");
-            Locale cookieLocale = org.orcid.jaxb.model.message.Locale.fromValue(clr.resolveLocale(request).toString());
-
-            // update the users preferences, so that
-            // send out emails in their last chosen language
-            if (!lastKnownLocale.equals(cookieLocale)) {
-                profileDao.updateLocale(orcidId, cookieLocale);
-            }
-        }
-    }
-    
-    private String determineFullTargetUrlFromSavedRequest(HttpServletRequest request, HttpServletResponse response) {
-        SavedRequest savedRequest = new HttpSessionRequestCache().getRequest(request, response);
-        String url = null;
-        if (savedRequest != null) {
-            url = savedRequest.getRedirectUrl();
-        }
-
-        // this next section is a hack, we should refactor to us
-        // some of kind of configuration file
-        String contextPath =  request.getContextPath();
-        if (url != null) {
-        	if (orcidUrlManager.getBasePath().equals("/") && !contextPath.equals("/"))
-        		url = url.replaceFirst(contextPath.replace("/", "\\/"), "");
-            if (url.contains("/signin/auth"))
-                url = null;
-            else if (url.contains(".json"))
-                url = null;
-        }
-
-        return url;
-    }
-
-    private String determineFullTargetUrl(HttpServletRequest request, HttpServletResponse response) {
-        return orcidUrlManager.getServerStringWithContextPath(request) + determineTargetUrl(request, response);
-    }
-    
     private String retrieveDisplayName(Map<String, String> headers) {
         String eppn = headers.get("eppn");
         if (StringUtils.isNotBlank(eppn)) {
@@ -215,5 +107,14 @@ public class ShibbolethAjaxAuthenticationSuccessHandler extends SimpleUrlAuthent
         }
         throw new OrcidBadRequestException("Couldn't find any user display name headers");
     }
-
+    
+    private String retrieveRemoteUser(Map<String, String> headers) {
+        for (String possibleHeader : POSSIBLE_REMOTE_USER_HEADERS) {
+            String userId = headers.get(possibleHeader);
+            if (userId != null) {
+                return userId;
+            }
+        }
+        throw new OrcidBadRequestException("Couldn't find remote user header");
+    }
 }
